backend/requirements.txt
fastapi>=0.110
uvicorn>=0.29
pymupdf>=1.23
faiss-cpu>=1.8
sentence-transformers>=2.6
google-genai>=1.7
aiosqlite>=0.20
python-multipart>=0.0.9
pydantic>=2.0
